Update Emacs versions to run on CI
In order to cut down the list a bit, I kept the latest minor of each series
since 25, and all the minor versions of the latest (currently 29) series. This
seems an ok rule for the future.

Also updates the Ruby version (it doesn't matter much or at all but 3.0 is old
already).